# src/app/api/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8082",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/configuracoes/v1/status-pedido/{id}": {
            "get": {
                "tags": [
                    "status-pedido-controller"
                ],
                "operationId": "obterPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/hal+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusPedidoResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "status-pedido-controller"
                ],
                "operationId": "atualizar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateStatusPedidoDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "status-pedido-controller"
                ],
                "operationId": "inativar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/configuracoes/v1/tiposervico": {
            "get": {
                "tags": [
                    "TipoServico"
                ],
                "summary": "GET method to handle operations for listaTipoServico",
                "description": "GET method to handle operations for listaTipoServico",
                "operationId": "listaTipoServico",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "codigo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categoriaId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pagina",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "0"
                        }
                    },
                    {
                        "name": "tamanho",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "20"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/WrapperListaTipoServicoDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "TipoServico"
                ],
                "summary": "POST method to handle operations for createTipoServico",
                "description": "POST method to handle operations for createTipoServico",
                "operationId": "createTipoServico",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CriarTiposServicosDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/configuracoes/v1/categoriaservico": {
            "get": {
                "tags": [
                    "CategoriaServico"
                ],
                "summary": "GET method to handle operations for listaCategoriaServico",
                "description": "GET method to handle operations for listaCategoriaServico",
                "operationId": "listaCategoriaServico",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "codigo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pagina",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "0"
                        }
                    },
                    {
                        "name": "tamanho",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "20"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/WrapperListaCategoriaServicoDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CategoriaServico"
                ],
                "summary": "POST method to handle operations for createCategoria",
                "description": "POST method to handle operations for createCategoria",
                "operationId": "createCategoria",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CriarCategoriasServicosDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CriarTiposServicosDTO": {
                "type": "object",
                "properties": {
                    "categoriaId": {
                        "type": "string"
                    },
                    "codigo": {
                        "type": "string"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "descricao": {
                        "type": "string"
                    },
                    "prazoEstimado": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "valorBase": {
                        "type": "number",
                        "format": "double"
                    },
                    "requerVistoria": {
                        "type": "boolean"
                    },
                    "requerAnaliseTec": {
                        "type": "boolean"
                    },
                    "requerAprovacao": {
                        "type": "boolean"
                    },
                    "disponivelPortal": {
                        "type": "boolean"
                    },
                    "ativo": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "codigo",
                    "nome"
                ]
            },
            "UpdateStatusPedidoDTO": {
                "type": "object",
                "properties": {
                    "codigo": {
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 2
                    },
                    "nome": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 3
                    },
                    "descricao": {
                        "type": "string"
                    },
                    "cor": {
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 0
                    },
                    "icone": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 0
                    },
                    "ordem": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "visivelPortal": {
                        "type": "boolean"
                    }
                }
            },
...
            "CriarCategoriasServicosDTO": {
                "type": "object",
                "properties": {
                    "nome": {
                        "type": "string"
                    },
                    "codigo": {
                        "type": "string"
                    },
                    "descricao": {
                        "type": "string"
                    },
                    "icone": {
                        "type": "string"
                    },
                    "cor": {
                        "type": "string"
                    },
                    "ordem": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "ativo": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "codigo",
                    "nome"
                ]
            },
...
            "ListaTipoServicoDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "tipoServicoId": {
                        "type": "string"
                    },
                    "codigo": {
                        "type": "string"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "categoria": {
                        "type": "string"
                    },
                    "estado": {
                        "type": "string"
                    }
                }
            },
            "WrapperListaTipoServicoDTO": {
                "type": "object",
                "properties": {
                    "pageNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ListaTipoServicoDTO"
                        }
                    }
                }
            },
...
            "StatusPedidoResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "codigo": {
                        "type": "string"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "descricao": {
                        "type": "string"
                    },
                    "cor": {
                        "type": "string"
                    },
                    "icone": {
                        "type": "string"
                    },
                    "ordem": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "visivelPortal": {
                        "type": "boolean"
                    }
                }
            },
            "ListaCategoriaDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "categoriaId": {
                        "type": "string"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "codigo": {
                        "type": "string"
                    },
                    "estado": {
                        "type": "string"
                    }
                }
            },
            "WrapperListaCategoriaServicoDTO": {
                "type": "object",
                "properties": {
                    "pageNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ListaCategoriaDTO"
                        }
                    }
                }
            }
        }
    }
}